Updated documentation
